refactor resolution to circular import
